+++ mpdb-client/src/edu/rpi/metpetdb/client/ui/image/browser/Point.java
@@ -2,7 +2,7 @@ package edu.rpi.metpetdb.client.ui.image.browser;
 
 public class Point {
 	
-	public int x = 0;
-	public int y = 0;
+	public double x = 0;
+	public double y = 0;
 
 }
