+++ mpdb-client/src/edu/rpi/metpetdb/client/ui/image/browser/click/listeners/RotateListener.java
@@ -29,8 +29,8 @@ public class RotateListener implements ClickListener {
 			public void onSuccess(final ImageOnGridContainer result) {
 				// final float widthRatio = iog.getWidth()
 				// / (float) iog.getImage().getWidth();
-				final float heightRatio = iog.getCurrentHeight()
-						/ (float) iog.getIog().getImage().getHeight();
+				final double heightRatio = iog.getCurrentHeight()
+						/  iog.getIog().getImage().getHeight();
 				iog.getIog().setImage(
 						(((ImageOnGridContainer) result).getIog().getImage()));
 				iog.getActualImage()
